#8 : move default directory
